requirements.txt
flask
pillow
pdf2image
pytesseract
reportlab
